Refactor D3DRenderHook
